moved organism options to interfaces
