+++ src/examples/sum/interfaces/sum-organism-options.ts
@@ -1,4 +1,4 @@
-import { IOrganismOptions } from "../../index";
+import { IOrganismOptions } from "../../../index";
 
 export interface ISumOrganismOptions extends IOrganismOptions {
     target: number;
